// backend/user_management/src/expert/dto/filterExpert.dto.ts
export class filterExpertDto {
    filter: {
        _id: any;
        email: any;
        address: any;
        tel: any;
        fullName: any;
        status: any;
        dateRange: any;
        repos: any;
        lat: any;
        lng: any;
        note: any;
    };
    specialite: any;
    pageNumber: number;
    pageSize: number;
    sortField: any;
    sortOrder: any;
}